// src/types/Types.ts
export interface UserInfo {
  nickname: string
  firstName: string
  lastName: string
  email: string
  phone: string
  github: string
  telegram: string
  resume: string
  advantages: string
  advantages2: string
  advantages3: string
  about: string
  file: File | null
  selectedSex: string[]
  checkbox1: boolean
  checkbox2: boolean
  checkbox3: boolean
  radio1: boolean
  radio2: boolean
  radio3: boolean
}